Make TimestampedLwM2mNodes immutable
