+++ leshan-core/src/main/java/org/eclipse/leshan/core/node/codec/senml/LwM2mNodeSenMLDecoder.java
@@ -42,7 +42,6 @@ import org.eclipse.leshan.core.node.LwM2mSingleResource;
 import org.eclipse.leshan.core.node.ObjectLink;
 import org.eclipse.leshan.core.node.TimestampedLwM2mNode;
 import org.eclipse.leshan.core.node.TimestampedLwM2mNodes;
-import org.eclipse.leshan.core.node.TimestampedLwM2mNodesImpl;
 import org.eclipse.leshan.core.node.codec.CodecException;
 import org.eclipse.leshan.core.node.codec.DefaultLwM2mDecoder;
 import org.eclipse.leshan.core.node.codec.MultiNodeDecoder;
@@ -186,7 +185,7 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
             // Decode SenML pack
             SenMLPack pack = decoder.fromSenML(content);
 
-            TimestampedLwM2mNodesImpl nodes = new TimestampedLwM2mNodesImpl();
+            TimestampedLwM2mNodes.Builder nodes = TimestampedLwM2mNodes.builder();
 
             LwM2mSenMLResolver resolver = new LwM2mSenMLResolver();
             for (SenMLRecord record : pack.getRecords()) {
@@ -194,10 +193,10 @@ public class LwM2mNodeSenMLDecoder implements TimestampedNodeDecoder, MultiNodeD
                 LwM2mPath path = resolvedRecord.getPath();
                 LwM2mNode node = parseRecords(Arrays.asList(resolvedRecord), path, model,
                         DefaultLwM2mDecoder.nodeClassFromPath(path));
-                nodes.add(new TimestampedLwM2mNodesImpl(resolvedRecord.getTimeStamp(), path, node));
+                nodes.put(resolvedRecord.getTimeStamp(), path, node);
             }
 
-            return nodes;
+            return nodes.build();
         } catch (SenMLException e) {
             String hexValue = content != null ? Hex.encodeHexString(content) : "";
             throw new CodecException(e, "Unable to decode nodes : %s", hexValue, e);
